refactored button so it can be added to side menu
caused break in cv code css
