+++ src/App.css
@@ -39,34 +39,25 @@
 .App {
   text-align: center;
 }
-.App-header {
-  background-color: #282c34;
-  background-image: url('./images/gridbook_pen_mess4.jpg');
-  background-size: cover;
-  background-repeat: no-repeat;
-  background-position: center;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: flex-start;
-  font-size: calc(10px + 5vmin);
+.header {
+  text-align: center;
+  font-size: calc(10px + 7vmin);
   color: midnightblue;
-  text-shadow: 2px 2px 0 white, -2px -2px 0 white, 2px -2px 0 white, -2px 2px 0 white; /* Create white outline */
-}
-.App-header h1 {
-  margin-top: 20px; /* Adjust the value as needed */
+  text-shadow: 2px 2px 0 white, -2px -2px 0 white, 2px -2px 0 white, -2px 2px 0 white;
 }
 
+.header h1 {
+  margin-top: 10%; /* Adjust the value as needed */
+}
 .home-bg {
   background-image: url('./images/gridbook_pen_mess5.jpg');
   background-size: cover;
   background-position: center;
   height: 100vh;
   display: flex;
-  flex-direction: column;
+  flex-direction: column; /* Change to column layout */
   justify-content: center;
-  align-items: flex-start;
+  align-items: center; /* Center horizontally */
 }
 
 .about-bg {
