+++ src/index.js
@@ -7,6 +7,8 @@ import About from './components/About';
 import CV from './components/CV';
 import Projects from './components/Projects';
 import Contact from './components/Contact';
+import Home from './components/Home';
+import PiperButton from './components/PiperButton';
 import reportWebVitals from './reportWebVitals';
 import { Link, BrowserRouter as Router, Route, Routes } from 'react-router-dom';
